Refactor TUI implementation with types package separation
- Create separate types package for shared type definitions (ClientInfo, StatsSummary, FullClientInfo)
- Move TUI implementation to dedicated tui package with proper modularization
- Resolve circular import issues between main and tui packages
- Update all API handlers and tests to use types package
- Clean up duplicate type definitions across codebase
- Improve package structure for better maintainability and test coverage
