+++ tui/format_test.go
@@ -1,4 +1,4 @@
-package trabbits_test
+package tui
 
 import (
 	"testing"
@@ -30,17 +30,9 @@ func TestFormatID(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			// We need to call the function via reflection or make it public
-			// For now, we'll test the functionality indirectly
-			if len(tt.input) <= 8 {
-				if tt.input != tt.expected {
-					t.Errorf("formatID(%q) = %q, want %q", tt.input, tt.input, tt.expected)
-				}
-			} else {
-				expected := tt.input[:6] + ".."
-				if expected != tt.expected {
-					t.Errorf("formatID(%q) = %q, want %q", tt.input, expected, tt.expected)
-				}
+			result := formatID(tt.input)
+			if result != tt.expected {
+				t.Errorf("formatID(%q) = %q, want %q", tt.input, result, tt.expected)
 			}
 		})
 	}
@@ -76,17 +68,7 @@ func TestFormatDuration(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			var result string
-			if tt.duration < time.Minute {
-				result = "30s ago"
-			} else if tt.duration < time.Hour {
-				result = "2m ago"
-			} else if tt.duration < 24*time.Hour {
-				result = "3h ago"
-			} else {
-				result = "1d ago"
-			}
-
+			result := formatDuration(tt.duration)
 			if result != tt.expected {
 				t.Errorf("formatDuration(%v) = %q, want %q", tt.duration, result, tt.expected)
 			}
@@ -124,17 +106,7 @@ func TestFormatNumber(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			var result string
-			if tt.input < 1000 {
-				result = "123"
-			} else if tt.input < 1000000 {
-				result = "1.5K"
-			} else if tt.input < 1000000000 {
-				result = "2.5M"
-			} else {
-				result = "3.5G"
-			}
-
+			result := formatNumber(tt.input)
 			if result != tt.expected {
 				t.Errorf("formatNumber(%d) = %q, want %q", tt.input, result, tt.expected)
 			}
@@ -171,30 +143,10 @@ func TestTruncate(t *testing.T) {
 
 	for _, tt := range tests {
 		t.Run(tt.name, func(t *testing.T) {
-			var result string
-			if len(tt.input) <= tt.maxLen {
-				result = tt.input
-			} else {
-				result = tt.input[:tt.maxLen-2] + ".."
-			}
-
+			result := truncate(tt.input, tt.maxLen)
 			if result != tt.expected {
 				t.Errorf("truncate(%q, %d) = %q, want %q", tt.input, tt.maxLen, result, tt.expected)
 			}
 		})
 	}
-}
-
-func TestTUIModelCreation(t *testing.T) {
-	// Test that we can create a TUI model without panicking
-	// This is a basic test since the actual TUI functionality requires a real terminal
-	defer func() {
-		if r := recover(); r != nil {
-			t.Errorf("TUI model creation panicked: %v", r)
-		}
-	}()
-
-	// We can't easily test the full TUI without a terminal environment
-	// but we can test that the basic structure works
-	t.Log("TUI model creation test passed")
 }
